backend/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2.19
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
freetype-py==2.3.0
idna==3.4
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pycairo==1.23.0
pycparser==2.21
PyJWT==2.7.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
reportlab==4.0.0
requests==2.30.0
requests-oauthlib==1.3.1
rlPyCairo==0.2.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
zipp==3.15.0
